+++ tests/Doctrine/Tests/Models/Translation/Article.php
@@ -54,6 +54,9 @@ class Article
     /** @PHPCRODM\String(translated=true, nullable=true) */
     public $nullable;
 
+    /** @PHPCRODM\String(translated=true, assoc="", nullable=true)*/
+    public $assoc;
+
     /**
      * @PHPCRODM\String(assoc="", translated=true, nullable=true)
      */